# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LumeScope API
  version: v1
  description: |
    LumeScope API provides access to Lumera blockchain action data, supernode metrics, 
    and version compatibility information. This API enables monitoring and analysis of 
    the Lumera network's distributed storage operations.
servers:
  - url: /
    description: Current server
paths:
...
  /v1/actions:
    get:
      summary: List actions
      description: Retrieve a paginated list of blockchain actions with optional filtering
      operationId: listActions
      tags:
        - Actions
      parameters:
        - name: type
          in: query
          description: Filter by action type
          schema:
            type: string
            enum:
              - ACTION_TYPE_UNSPECIFIED
              - ACTION_TYPE_CASCADE
              - ACTION_TYPE_SENSE
        - name: creator
          in: query
          description: Filter by creator address
          schema:
            type: string
        - name: state
          in: query
          description: Filter by action state
          schema:
            type: string
            enum:
              - ACTION_STATE_UNSPECIFIED
              - ACTION_STATE_PENDING
              - ACTION_STATE_DONE
              - ACTION_STATE_APPROVED
              - ACTION_STATE_EXPIRED
        - name: from
          in: query
          description: Filter actions from block height
          schema:
            type: integer
            format: int64
        - name: to
          in: query
          description: Filter actions to block height
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          description: Maximum number of items to return (1-200)
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          description: Pagination cursor for next page
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionsListResponse'
  /v1/actions/{id}:
    get:
      summary: Get action details
      description: Retrieve detailed information about a specific action by ID
      operationId: getAction
      tags:
        - Actions
      parameters:
        - name: id
          in: path
          required: true
          description: Action ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionDetail'
        '404':
          description: Action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/actions/stats:
    get:
      summary: Get global action statistics
      description: |
        Retrieve aggregated action statistics for all actions (global).
        Returns total count and a breakdown by state (PENDING, DONE, APPROVED, EXPIRED).
      operationId: getActionStats
      tags:
        - Actions
      parameters:
        - name: type
          in: query
          description: Filter by action type
          schema:
            type: string
            enum:
              - ACTION_TYPE_UNSPECIFIED
              - ACTION_TYPE_CASCADE
              - ACTION_TYPE_SENSE
      responses:
        '200':
          description: Successful response with action statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionStatsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    ActionStatsResponse:
      type: object
      description: Aggregated action statistics for all actions (global)
      properties:
        total:
          type: integer
          description: Total number of actions
          example: 100
        states:
          type: object
          description: Count of actions grouped by state
          additionalProperties:
            type: integer
          example:
            ACTION_STATE_PENDING: 10
            ACTION_STATE_DONE: 80
            ACTION_STATE_APPROVED: 5
            ACTION_STATE_EXPIRED: 5
        schema_version:
          type: string
          description: API schema version
          example: v1.0
      required:
        - total
        - states
        - schema_version
...
    ActionsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ActionItem'
        next_cursor:
          type: string
          description: Cursor for fetching the next page
        schema_version:
          type: string
          example: v1.0
    ActionItem:
      type: object
      properties:
        id:
          type: string
          description: Unique action identifier
          example: act_1
        type:
          type: string
          description: Action type (e.g., cascade, transfer)
          example: cascade
        creator:
          type: string
          description: Creator address
          example: lumera1example...
        state:
          type: string
          description: Current action state
          example: committed
        block_height:
          type: integer
          format: int64
          description: Block height when action was created
          example: 12345
        decoded:
          type: object
          additionalProperties: true
          description: Decoded action data
        raw:
          type: string
          description: Base64 encoded raw bytes for unknown types
        superNodes:
          type: array
          description: List of SuperNode operator addresses associated with the action
          items:
            type: string
            example: lumera1563uuzljqvpanh79w2tvymqsly73v9nwygda73
    ActionDetail:
      type: object
      properties:
        id:
          type: string
          description: Unique action identifier
        type:
          type: string
          description: Action type
          example: cascade
        creator:
          type: string
          description: Creator address
        state:
          type: string
          description: Current action state
          example: committed
        block_height:
          type: integer
          format: int64
          description: Block height when action was created
          example: 12345
        price:
          $ref: '#/components/schemas/Price'
        timestamp:
          type: string
          format: date-time
          description: Action timestamp
        decoded:
          type: object
          additionalProperties: true
          description: Decoded action data
        raw:
          type: string
          description: Base64 encoded raw bytes for unknown types
        super_nodes:
          type: object
          additionalProperties: true
          description: SuperNode data associated with the action
        schema_version:
          type: string
          example: v1.0
    Price:
      type: object
      description: Price information for an action
      properties:
        denom:
          type: string
          description: Price denomination
          example: ulume
        amount:
          type: string
          description: Price amount as string
          example: "1000000"
...
    Error:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error message
          example: internal_error
      required:
        - error
...